First-stage refactoring of the block mesh renderer: Breaking the renderer into distinct parts that are overridable on a block-by-block basis. Not yet definable via the block json, that will be a later step once a few other things are complete first.
